# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitnes Planer API",
        "description": "Dokumentacija koristeći PHP Atribute",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Lokalni API server"
        }
    ],
    "paths": {
...
        "/programi": {
            "get": {
                "tags": [
                    "Programi"
                ],
                "summary": "Prikaz dostupnih programa (javni + sopstveni)",
                "operationId": "6e1452c63cc0c5af4a11eba39ccac6ca",
                "responses": {
                    "200": {
                        "description": "Lista programa"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Programi"
                ],
                "summary": "Kreiranje novog programa sa vežbama",
                "operationId": "f13ae3fe1ab50a6f3dd39f866f0b1261",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "vezbe",
                                    "public"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string"
                                    },
                                    "public": {
                                        "type": "boolean"
                                    },
                                    "vezbe": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "dan": {
                                                    "type": "integer"
                                                },
                                                "serija": {
                                                    "type": "integer"
                                                },
                                                "ponavljanja": {
                                                    "type": "integer"
                                                },
                                                "tezina": {
                                                    "type": "number"
                                                },
                                                "trajanje": {
                                                    "type": "integer"
                                                },
                                                "bpm": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Program kreiran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/programi/{id}": {
            "put": {
                "tags": [
                    "Programi"
                ],
                "summary": "Ažuriranje programa",
                "operationId": "870bae7a628f827a545398072c5cc9c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Program ažuriran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Programi"
                ],
                "summary": "Brisanje sopstvenog programa",
                "operationId": "4af69d47b4cf465d566e2500083a77f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/programi/treneri": {
            "get": {
                "tags": [
                    "Programi"
                ],
                "summary": "Programi trenera (zavisi da li korisnik ima trenera)",
                "operationId": "16f4424ad2f5b18d528458137cbf95da",
                "responses": {
                    "200": {
                        "description": "Lista programa"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/programi/moji": {
            "get": {
                "tags": [
                    "Programi"
                ],
                "summary": "Prikaz svih programa koje je kreirao ulogovani korisnik",
                "operationId": "e9647c5141e655c508cf7a275ca0d3fe",
                "responses": {
                    "200": {
                        "description": "Lista mojih programa"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/programi/dodaj-veze": {
            "post": {
                "tags": [
                    "Programi"
                ],
                "summary": "Brzo kreiranje 'Trening trenera' programa",
                "operationId": "0a5b69773517f39dca40a8a09165f3a6",
                "responses": {
                    "201": {
                        "description": "Program dodat"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}